requirements.txt
apiclient==1.0.3
asn1crypto==0.24.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-vision==0.35.1
googleapis-common-protos==1.5.5
grpcio==1.17.0
httplib2==0.12.0
idna==2.8
Markdown==3.0.1
oauth2client==4.1.3
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
